# docs/phase1_openapi.yaml
openapi: 3.0.3
info:
  title: 智享营销系统第 1 阶段 API
  version: 1.0.0
  description: >
    第 1 阶段 MVP 接口定义，覆盖后台管理、门店端、小程序端、微信支付、销售单微信分享收款。
    价格、库存、订单状态、客户身份、收款状态均由后端裁剪和流转，前端不得自行判断。
servers:
  - url: https://api.example.com
    description: 生产环境占位
  - url: http://localhost:8080
    description: 本地开发环境
...
security:
  - bearerAuth: []
paths:
  /api/admin/auth/login:
    post:
      tags: [AdminAuth]
      summary: 后台或门店端登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
...
  /api/admin/stores:
    get:
      tags: [AdminStore]
      summary: 门店列表
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 门店分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorePageResponse'
    post:
      tags: [AdminStore]
      summary: 新增门店
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreUpsertRequest'
      responses:
        '200':
          description: 新增成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
  /api/admin/products:
    get:
      tags: [AdminProduct]
      summary: 后台商品列表
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: keyword
          in: query
          description: 商品名称、SKU编码、条码
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, ON_SALE, OFF_SALE]
        - name: categoryId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 商品分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPageResponse'
    post:
      tags: [AdminProduct]
      summary: 新增商品和 SKU
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '200':
          description: 新增成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
  /api/admin/products/{skuId}/price:
    put:
      tags: [AdminProduct]
      summary: 修改 SKU 价格
      description: 修改零售价、批发价、小程序价或门店价；批发价修改必须写价格日志。
      parameters:
        - $ref: '#/components/parameters/SkuId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductPriceUpdateRequest'
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/admin/reports/dashboard:
    get:
      tags: [AdminReport]
      summary: 后台首页经营看板
      parameters:
        - name: storeId
          in: query
          schema:
            type: integer
            format: int64
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 首页经营数据
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DashboardSummary'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    miniappAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SkuId:
      name: skuId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "0"
        message:
          type: string
          example: 成功
        traceId:
          type: string
          example: 01HZTRACEID
    PageResponse:
      type: object
      required: [total, page, pageSize, records]
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        records:
          type: array
          items:
            type: object
    IdResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        realName:
          type: string
        storeId:
          type: integer
          format: int64
          nullable: true
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
    StoreUpsertRequest:
      type: object
      required: [name, address, contact, phone]
      properties:
        name:
          type: string
        address:
          type: string
        lng:
          type: number
          format: double
        lat:
          type: number
          format: double
        contact:
          type: string
        phone:
          type: string
        deliveryRadius:
          type: number
          format: double
    Store:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        businessStatus:
          type: string
          enum: [OPEN, PAUSED, CLOSED]
        deliveryRadius:
          type: number
    StorePageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageResponse'
                - type: object
                  properties:
                    records:
                      type: array
                      items:
                        $ref: '#/components/schemas/Store'
    ProductCreateRequest:
      type: object
      required: [name, categoryId, saleChannels, skus]
      properties:
        name:
          type: string
        categoryId:
          type: integer
          format: int64
        mainImage:
          type: string
        saleChannels:
          type: array
          items:
            type: string
            enum: [MINIAPP, STORE]
        skus:
          type: array
          items:
            $ref: '#/components/schemas/SkuCreateRequest'
    SkuCreateRequest:
      type: object
      required: [skuName, boxRatio, retailPrice]
      properties:
        skuName:
          type: string
        barcode:
          type: string
        boxRatio:
          type: integer
        temperature:
          type: string
          enum: [NORMAL, CHILLED]
        traceEnabled:
          type: boolean
        warningThreshold:
          type: integer
        costPrice:
          type: number
        retailPrice:
          type: number
        wholesalePrice:
          type: number
          nullable: true
        miniappPrice:
          type: number
          nullable: true
        storePrice:
          type: number
          nullable: true
    ProductPriceUpdateRequest:
      type: object
      properties:
        costPrice:
          type: number
        retailPrice:
          type: number
        wholesalePrice:
          type: number
          nullable: true
        miniappPrice:
          type: number
          nullable: true
        storePrice:
          type: number
          nullable: true
    ProductItem:
      type: object
      properties:
        spuId:
          type: integer
          format: int64
        skuId:
          type: integer
          format: int64
        name:
          type: string
        skuName:
          type: string
        skuCode:
          type: string
        barcode:
          type: string
        retailPrice:
          type: number
        wholesalePrice:
          type: number
          nullable: true
        status:
          type: string
    ProductPageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageResponse'
                - type: object
                  properties:
                    records:
                      type: array
                      items:
                        $ref: '#/components/schemas/ProductItem'
...
    DashboardSummary:
      type: object
      properties:
        salesAmount:
          type: number
        orderCount:
          type: integer
        saleBillCount:
          type: integer
        pendingCollectionAmount:
          type: number
        inventoryWarningCount:
          type: integer
        pendingOrderCount:
          type: integer